+++ addons/filamind_iot_proxy_admin/models/__init__.py
@@ -1,0 +1,4 @@
+from . import proxy_config
+from . import proxy_tenant
+from . import proxy_box
+from . import proxy_audit